IMPORTANT: removed ProcessedEventPurger dependencies
