+++ samples/BoltOn.Samples.Application/Handlers/CreateStudentHandler.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using BoltOn.Cqrs;
 using BoltOn.Data;
 using BoltOn.Mediator.Pipeline;
 using BoltOn.Samples.Application.Entities;
@@ -33,27 +32,7 @@ namespace BoltOn.Samples.Application.Handlers
 					request.FirstName,
 					request.LastName
 				), cancellationToken);
-			System.Console.WriteLine("message from CreateStudentHandler: " + request.FirstName);
-		}
-	}
-
-	public class StudentCreatedEvent : EventToBeProcessed
-	{
-	}
-
-	public class StudentCreatedEventHandler : IRequestAsyncHandler<StudentCreatedEvent>
-	{
-		private readonly IRepository<StudentFlattened> _repository;
-
-		public StudentCreatedEventHandler(IRepository<StudentFlattened> repository)
-		{
-			_repository = repository;
-		}
-
-		public async Task HandleAsync(StudentCreatedEvent request, CancellationToken cancellationToken)
-		{
-			await _repository.AddAsync(new StudentFlattened(request), cancellationToken);
-			//await Task.CompletedTask;
+			Console.WriteLine("message from CreateStudentHandler: " + request.FirstName);
 		}
 	}
 }
